Add a new file and rename VideoStreamHandler to VideoDatagramAssembler
